+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Please use 2.3 if you need to run ``doc2dash`` on legacy Python versions.
 
 - ``zope.interface`` is not a dependency anymore.
 - ``colorama`` is only a dependency on Windows now.
+- Work around a display bug in older Dash.app releases by setting the plist key ``DashDocSetDeclaredInStyle`` to ``originalName``.
 
 
 ----